[feat] NetworkModule, ServiceModule 분리
