+++ app/src/main/java/com/eatssu/android/di/NetworkModule.kt
@@ -3,12 +3,7 @@ package com.eatssu.android.di
 
 import com.eatssu.android.BuildConfig
 import com.eatssu.android.BuildConfig.BASE_URL
-import com.eatssu.android.data.service.MealService
-import com.eatssu.android.data.service.MenuService
 import com.eatssu.android.data.service.OauthService
-import com.eatssu.android.data.service.ReportService
-import com.eatssu.android.data.service.ReviewService
-import com.eatssu.android.data.service.UserService
 import com.eatssu.android.di.network.TokenInterceptor
 import dagger.Module
 import dagger.Provides
@@ -72,35 +67,5 @@ object NetworkModule {
     @Singleton
     fun provideOauthService(retrofit: Retrofit): OauthService {
         return retrofit.create(OauthService::class.java)
-    }
-
-    @Provides
-    @Singleton
-    fun provideUserService(retrofit: Retrofit): UserService {
-        return retrofit.create(UserService::class.java)
-    }
-
-    @Provides
-    @Singleton
-    fun provideReportService(retrofit: Retrofit): ReportService {
-        return retrofit.create(ReportService::class.java)
-    }
-
-    @Provides
-    @Singleton
-    fun provideReviewService(retrofit: Retrofit): ReviewService {
-        return retrofit.create(ReviewService::class.java)
-    }
-
-    @Provides
-    @Singleton
-    fun provideMealService(retrofit: Retrofit): MealService {
-        return retrofit.create(MealService::class.java)
-    }
-
-    @Provides
-    @Singleton
-    fun provideMenuService(retrofit: Retrofit): MenuService {
-        return retrofit.create(MenuService::class.java)
     }
 }
